LayoutTests/webaudio/convolver-setBuffer-null-expected.txt

Tests that ConvolverNode impulse response buffer can be set to null.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS conv.buffer = null did not throw exception.
PASS conv.buffer is null
PASS conv.buffer = document threw exception TypeError: The ConvolverNode.buffer attribute must be an instance of AudioBuffer.
PASS conv.buffer is null
PASS successfullyParsed is true

TEST COMPLETE
